+++ netcode/netcode.c
@@ -126,7 +126,7 @@ void netcode_printf( int level, NETCODE_CONST char * format, ... )
 
 #endif // #if NETCODE_ENABLE_LOGGING
 
-void * netcode_default_allocate_function( void * context, uint64_t bytes )
+void * netcode_default_allocate_function( void * context, size_t bytes )
 {
     (void) context;
     return malloc( bytes );
@@ -581,9 +581,9 @@ void netcode_term()
 // ----------------------------------------------------------------
 
 #if NETCODE_PLATFORM == NETCODE_PLATFORM_WINDOWS
-typedef uint64_t netcode_socket_handle_t;
+typedef uint32_t netcode_socket_handle_t;
 #else // #if NETCODE_PLATFORM == NETCODE_PLATFORM_WINDOWS
-typedef int netcode_socket_handle_t;
+typedef size_t netcode_socket_handle_t;
 #endif // #if NETCODE_PLATFORM == NETCODe_PLATFORM_WINDOWS
 
 struct netcode_socket_t
@@ -591,7 +591,6 @@ struct netcode_socket_t
     struct netcode_address_t address;
     netcode_socket_handle_t handle;
 };
-
 
 struct netcode_socket_holder_t
 {
@@ -1022,16 +1021,16 @@ void netcode_random_bytes( uint8_t * data, int bytes )
 }
 
 int netcode_encrypt_aead_bignonce( uint8_t * message, uint64_t message_length, 
-                          uint8_t * additional, uint64_t additional_length,
-                          NETCODE_CONST uint8_t * nonce,
-                          NETCODE_CONST uint8_t * key )
+                                   uint8_t * additional, uint64_t additional_length,
+                                   NETCODE_CONST uint8_t * nonce,
+                                   NETCODE_CONST uint8_t * key )
 {
     unsigned long long encrypted_length;
 
     int result = crypto_aead_xchacha20poly1305_ietf_encrypt( message, &encrypted_length,
-                                                            message, (unsigned long long) message_length,
-                                                            additional, (unsigned long long) additional_length,
-                                                            NULL, nonce, key );
+                                                             message, (unsigned long long) message_length,
+                                                             additional, (unsigned long long) additional_length,
+                                                             NULL, nonce, key );
     
     if ( result != 0 )
         return NETCODE_ERROR;
@@ -1042,17 +1041,17 @@ int netcode_encrypt_aead_bignonce( uint8_t * message, uint64_t message_length,
 }
 
 int netcode_decrypt_aead_bignonce( uint8_t * message, uint64_t message_length, 
-                          uint8_t * additional, uint64_t additional_length,
-                          uint8_t * nonce,
-                          uint8_t * key )
+                                   uint8_t * additional, uint64_t additional_length,
+                                   uint8_t * nonce,
+                                   uint8_t * key )
 {
     unsigned long long decrypted_length;
 
     int result = crypto_aead_xchacha20poly1305_ietf_decrypt( message, &decrypted_length,
-                                                            NULL,
-                                                            message, (unsigned long long) message_length,
-                                                            additional, (unsigned long long) additional_length,
-                                                            nonce, key );
+                                                             NULL,
+                                                             message, (unsigned long long) message_length,
+                                                             additional, (unsigned long long) additional_length,
+                                                             nonce, key );
 
     if ( result != 0 )
         return NETCODE_ERROR;
@@ -1462,7 +1461,7 @@ struct netcode_connection_disconnect_packet_t
     uint8_t packet_type;
 };
 
-struct netcode_connection_payload_packet_t * netcode_create_payload_packet( int payload_bytes, void * allocator_context, void* (*allocate_function)(void*,uint64_t) )
+struct netcode_connection_payload_packet_t * netcode_create_payload_packet( int payload_bytes, void * allocator_context, void* (*allocate_function)(void*,size_t) )
 {
     netcode_assert( payload_bytes >= 0 );
     netcode_assert( payload_bytes <= NETCODE_MAX_PAYLOAD_BYTES );
@@ -1710,7 +1709,7 @@ void * netcode_read_packet( uint8_t * buffer,
                             uint8_t * allowed_packets, 
                             struct netcode_replay_protection_t * replay_protection, 
                             void * allocator_context, 
-                            void* (*allocate_function)(void*,uint64_t) )
+                            void* (*allocate_function)(void*,size_t) )
 {
     netcode_assert( sequence );
     netcode_assert( allowed_packets );
@@ -2267,7 +2266,7 @@ int netcode_read_connect_token( uint8_t * buffer, int buffer_length, struct netc
 struct netcode_packet_queue_t
 {
     void * allocator_context;
-    void * (*allocate_function)(void*,uint64_t);
+    void * (*allocate_function)(void*,size_t);
     void (*free_function)(void*,void*);
     int num_packets;
     int start_index;
@@ -2277,7 +2276,7 @@ struct netcode_packet_queue_t
 
 void netcode_packet_queue_init( struct netcode_packet_queue_t * queue, 
                                 void * allocator_context, 
-                                void * (*allocate_function)(void*,uint64_t), 
+                                void * (*allocate_function)(void*,size_t), 
                                 void (*free_function)(void*,void*) )
 {
     if ( allocate_function == NULL )
@@ -2359,7 +2358,7 @@ struct netcode_network_simulator_packet_entry_t
 struct netcode_network_simulator_t
 {
     void * allocator_context;
-    void * (*allocate_function)(void*,uint64_t);
+    void * (*allocate_function)(void*,size_t);
     void (*free_function)(void*,void*);
     float latency_milliseconds;
     float jitter_milliseconds;
@@ -2373,7 +2372,7 @@ struct netcode_network_simulator_t
 };
 
 struct netcode_network_simulator_t * netcode_network_simulator_create( void * allocator_context, 
-                                                                       void * (*allocate_function)(void*,uint64_t), 
+                                                                       void * (*allocate_function)(void*,size_t), 
                                                                        void (*free_function)(void*,void*) )
 {
     if ( allocate_function == NULL )
